# perf: avoid an extra copy for otlp protobuf payload during unmarshaling