--- a/route/route.go
+++ b/route/route.go
@@ -643,7 +643,11 @@ func (router *Router) processOTLPRequestBatchMsgp(
 		totalEvents += len(batch.Events)
 		for _, ev := range batch.Events {
 			payload := types.NewPayload(router.Config, nil)
-			coreFieldsUnmarshaler.UnmarshalPayload(ev.Attributes, &payload)
+			err := coreFieldsUnmarshaler.UnmarshalPayloadComplete(ev.Attributes, &payload)
+			if err != nil {
+				router.Logger.Error().Logf("Error unmarshaling payload: " + err.Error())
+				continue
+			}
 
 			event := &types.Event{
 				Context:     ctx,